+++ kv/etcd/client.go
@@ -16,8 +16,8 @@ import (
 	"github.com/influxdata/chronograf/mocks"
 	"github.com/influxdata/chronograf/server/config"
 	"github.com/influxdata/chronograf/snowflake"
-	"go.etcd.io/etcd/clientv3"
-	"go.etcd.io/etcd/clientv3/concurrency"
+	clientv3 "go.etcd.io/etcd/client/v3"
+	"go.etcd.io/etcd/client/v3/concurrency"
 )
 
 const (