+++ ui/src/kapacitor/constants/index.ts
@@ -33,6 +33,7 @@ export enum AlertTypes {
   log = 'log',
   servicenow = 'servicenow',
   bigpanda = 'bigpanda',
+  teams = 'teams',
 }
 
 export enum AlertDisplayText {
@@ -51,6 +52,7 @@ export enum AlertDisplayText {
   victorops = 'VictorOps',
   servicenow = 'ServiceNow',
   bigpanda = 'BigPanda',
+  teams = 'Teams',
 }
 
 export const SupportedServices: string[] = [
@@ -67,6 +69,7 @@ export const SupportedServices: string[] = [
   'slack',
   'smtp',
   'talk',
+  'teams',
   'telegram',
   'victorops',
 ]
@@ -215,6 +218,7 @@ export const MAP_KEYS_FROM_CONFIG: KeyMappings = {
   victorops: 'victorOps',
   servicenow: 'serviceNow',
   bigpanda: 'bigPanda',
+  teams: 'teams',
 }
 
 // ALERTS_FROM_CONFIG the array of fields to accept from Kapacitor Config
@@ -243,6 +247,9 @@ export const ALERTS_FROM_CONFIG: FieldsFromConfigAlerts = {
   // mqtt:[]
   serviceNow: ['url', 'source', 'username', 'password'],
   bigPanda: ['app-key'],
+  teams: [
+    /* 'channel-url' is intentionally skipped */
+  ],
 }
 
 export const MAP_FIELD_KEYS_FROM_CONFIG: ConfigKeyMaps = {
@@ -268,6 +275,7 @@ export const MAP_FIELD_KEYS_FROM_CONFIG: ConfigKeyMaps = {
   // mqtt: {}
   serviceNow: {},
   bigPanda: {},
+  teams: {},
 }
 
 // HANDLERS_TO_RULE returns array of fields that may be updated for each alert on rule.
@@ -312,4 +320,5 @@ export const HANDLERS_TO_RULE_THEM_ALL: FieldsFromAllAlerts = {
     '_type', // serviceNow type, remapped from type on the wire
   ],
   bigPanda: ['app-key', 'primary-property', 'secondary-property'],
+  teams: ['channel-url'],
 }